+++ tests/unit/test_transformer.py
@@ -60,8 +60,7 @@ def transformer(sagemaker_session):
 
 
 def test_delete_model(sagemaker_session):
-    transformer = Transformer(MODEL_NAME, INSTANCE_COUNT, INSTANCE_TYPE, output_path=OUTPUT_PATH,
-                              sagemaker_session=sagemaker_session, volume_kms_key=KMS_KEY_ID)
+    transformer = Transformer(MODEL_NAME, INSTANCE_COUNT, INSTANCE_TYPE, sagemaker_session=sagemaker_session)
     transformer.delete_model()
     sagemaker_session.delete_model.assert_called_with(MODEL_NAME)